// src/app/models/aetPostos.ts
export interface AetPostos{
    id?:                       any;
    cod?:                   String;
    cargo?:                    any;
    fatoresRisco?:          String;
    grauRisco?:             String;
    segCorpoPrinc?:            any;
    demSegm?:                  any;
    diagnosticoGlobal?:     String;
    recomendacoes?:         String;
    nomeCargo?:             String;
    nomeSegCorpoPrinc?:     String;
    nomeDemSegm?:           String;
}